Added float FU and load immediate float, also lab4c/the loop

+++ include/register_file.h
@@ -6,6 +6,6 @@ class RegisterFile {
 	public:
 		RegisterFile();
 		int ir;
-		map<string, int> registers;
+		map<string, float> registers;
 
 };
